requirements.txt
pyodbc==5.2.0
PyQt5==5.15.9
python-dotenv==1.0.0
